# Misc/CastleSwaggerApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo API",
    "version": "v1"
  },
  "paths": {
    "/c2profile": {
      "post": {
        "tags": [
          "C2Profile"
        ],
        "operationId": "SetProfile",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "C2Profile"
        ],
        "operationId": "GetCurrentProfile",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              }
            }
          }
        }
      }
    },
...
    "/c2profile/{name}": {
      "get": {
        "tags": [
          "C2Profile"
        ],
        "operationId": "GetProfile",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "C2Profile"
        ],
        "operationId": "DeleteProfile",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/c2profile/create": {
      "post": {
        "tags": [
          "C2Profile"
        ],
        "operationId": "CreateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateC2ProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateC2ProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateC2ProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              }
            }
          }
        }
      }
    },
    "/c2profile/update": {
      "post": {
        "tags": [
          "C2Profile"
        ],
        "operationId": "UpdateProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateC2ProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateC2ProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateC2ProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2Profile"
                }
              }
            }
          }
        }
      }
    },
    "/commander": {
      "get": {
        "tags": [
          "Commander"
        ],
        "operationId": "GetCommanders",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Commander"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Commander"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Commander"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Commander"
        ],
        "operationId": "StartCommander",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartHttpCommanderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartHttpCommanderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartHttpCommanderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              }
            }
          }
        }
      }
    },
    "/commander/{name}": {
      "get": {
        "tags": [
          "Commander"
        ],
        "operationId": "GetCommander",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Commander"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Commander"
        ],
        "operationId": "StopCommander",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "C2Profile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "http": {
            "$ref": "#/components/schemas/HttpProfile"
          }
        },
        "additionalProperties": false
      },
      "COMMANDER_TYPE": {
        "enum": [
          0
        ],
        "type": "integer",
        "format": "int32"
      },
      "Commander": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/COMMANDER_TYPE"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "bindPort": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "tls": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CreateC2ProfileRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "sleep": {
            "type": "integer",
            "format": "int32"
          },
          "jitter": {
            "type": "integer",
            "format": "int32"
          },
          "passwd": {
            "type": "string",
            "nullable": true
          },
          "getPaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "postPaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HttpProfile": {
        "type": "object",
        "properties": {
          "sleep": {
            "type": "integer",
            "format": "int32"
          },
          "jitter": {
            "type": "integer",
            "format": "int32"
          },
          "passwd": {
            "type": "string",
            "nullable": true
          },
          "getPaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "postPaths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StartHttpCommanderRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "bindPort": {
            "type": "integer",
            "format": "int32"
          },
          "tls": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}